attack/defends complete reworking game and ui. Didn't handle fight properly. Still need to add testing

+++ src/main/java/Character.java
@@ -40,7 +40,7 @@ public class Character {
     }
 
     public void defend(int attack) {
-        health = health - defendStrategy.defend(attack);
+        health = (int) (health - defendStrategy.defend(attack));
     }
 
     public void displayStats() {
